style(api): Saperate main server code and export code
